requirements/base.txt
psycopg2==2.5.1
Pillow==2.3.0
Django==1.6.1
South==0.8.4
-e git://github.com/vkurup/django_compressor.git@develop#egg=django_compressor
# Required by django-compressor
django-appconf==0.6
six==1.4.1

BeautifulSoup4==4.3.2
django-tastypie==0.11.0
